reservation canceled status and function and implement user limit data in all reservation functions

+++ src/main/java/com/hesho/reservation/web/rest/ReservationResource.java
@@ -1,5 +1,6 @@
 package com.hesho.reservation.web.rest;
 
+import com.hesho.reservation.domain.Reservation;
 import com.hesho.reservation.domain.User;
 import com.hesho.reservation.domain.enumeration.ReservationStatus;
 import com.hesho.reservation.security.AuthoritiesConstants;
@@ -97,7 +98,11 @@ public class ReservationResource {
             throw new BadRequestAlertException("No user detected", ENTITY_NAME, "noUser");
         }
         if (reservationDTO.getId() == null) {
-            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+            throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idNull");
+        }
+        Optional<ReservationDTO> oldReservation = reservationService.findOne(reservationDTO.getId(),user.get());
+        if (!oldReservation.isPresent()) {
+            throw new BadRequestAlertException("Reservation Not Registered with User", ENTITY_NAME, "NoReservation");
         }
         reservationDTO.setUserId(user.get().getId());
         ReservationDTO result = reservationService.save(reservationDTO);
@@ -118,6 +123,29 @@ public class ReservationResource {
         log.debug("REST request to update Reservation Status: {}", id);
 
         ReservationDTO result = reservationService.updateStatus(id,status);
+        return ResponseEntity.ok().body(result);
+    }
+
+    /**
+     * {@code PUT  /reservations} : Cancel an existing status reservation.
+     *
+     * @return the {@link ResponseEntity} with status {@code 200 (OK)} and with body the updated reservationDTO,
+     * or with status {@code 400 (Bad Request)} if the reservationDTO is not valid,
+     * or with status {@code 500 (Internal Server Error)} if the reservationDTO couldn't be updated.
+     * @throws URISyntaxException if the Location URI syntax is incorrect.
+     */
+    @PutMapping("/reservations/canceled/{id}")
+    public ResponseEntity<ReservationDTO> cancelReservation(@PathVariable Long id) throws URISyntaxException {
+        log.debug("REST request to cancel Reservation Status: {}", id);
+        Optional<User> user = userService.findCurrentUser();
+        if (!user.isPresent()) {
+            throw new BadRequestAlertException("No user detected", ENTITY_NAME, "noUser");
+        }
+        Optional<ReservationDTO> reservationDTO = reservationService.findOne(id,user.get());
+        if (!reservationDTO.isPresent()) {
+            throw new BadRequestAlertException("No Reservation detected", ENTITY_NAME, "NoReservation");
+        }
+        ReservationDTO result = reservationService.cancelReservation(id);
         return ResponseEntity.ok().body(result);
     }
 
@@ -155,7 +183,11 @@ public class ReservationResource {
     @GetMapping("/reservations/{id}")
     public ResponseEntity<ReservationDTO> getReservation(@PathVariable Long id) {
         log.debug("REST request to get Reservation : {}", id);
-        Optional<ReservationDTO> reservationDTO = reservationService.findOne(id);
+        Optional<User> user = userService.findCurrentUser();
+        if (!user.isPresent()) {
+            throw new BadRequestAlertException("No user detected", ENTITY_NAME, "noUser");
+        }
+        Optional<ReservationDTO> reservationDTO = reservationService.findOne(id,user.get());
         return ResponseUtil.wrapOrNotFound(reservationDTO);
     }
     /**
@@ -167,6 +199,14 @@ public class ReservationResource {
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         log.debug("REST request to delete Reservation : {}", id);
+        Optional<User> user = userService.findCurrentUser();
+        if (!user.isPresent()) {
+            throw new BadRequestAlertException("No user detected", ENTITY_NAME, "noUser");
+        }
+        Optional<ReservationDTO> reservationDTO = reservationService.findOne(id,user.get());
+        if (!reservationDTO.isPresent()) {
+            throw new BadRequestAlertException("No Reservation detected", ENTITY_NAME, "NoReservation");
+        }
         reservationService.delete(id);
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString())).build();
     }
